+++ site/components/illustrations/patterns/forms/overview/hero.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {getSSRId} from 'newskit';
-import {Svg} from '../../svg';
-import {Path} from '../../path';
+import {Svg} from '../../../svg';
+import {Path} from '../../../path';
 
 export const Hero: React.FC = () => {
   const mask0 = getSSRId();
